+++ composer/src/main/java/io/piano/android/composer/ComposerException.kt
@@ -1,8 +1,0 @@
-package io.piano.android.composer
-
-class ComposerException : RuntimeException {
-    constructor() {}
-    constructor(message: String?) : super(message) {}
-    constructor(message: String?, cause: Throwable?) : super(message, cause) {}
-    constructor(cause: Throwable?) : super(cause) {}
-}