+++ composer/src/main/java/io/piano/android/composer/DependenciesProvider.kt
@@ -25,6 +25,7 @@ internal class DependenciesProvider private constructor(
         .readTimeout(30, TimeUnit.SECONDS)
         .writeTimeout(30, TimeUnit.SECONDS)
         .addInterceptor(UserAgentInterceptor(userAgent))
+        .addInterceptor(RequestPolicyInterceptor(prefsStorage))
         .addInterceptor(
             HttpLoggingInterceptor().setLevel(
                 if (BuildConfig.DEBUG || isLogHttpSet())