+++ plugins/src/main/kotlin/io/piano/android/dependencies/Deps.kt
@@ -8,14 +8,14 @@ private object Versions {
     const val androidxActivity = "1.4.0"
     const val annotationsLibrary = "1.3.0"
     const val fragmentLibrary = "1.4.1"
-    const val materialLibrary = "1.5.0"
+    const val materialLibrary = "1.6.0"
     const val lifecycleLibrary = "2.4.1"
     const val prefsLibrary = "1.2.0"
     const val multidex = "2.0.1"
 
     // Third party Libraries
-    const val facebookLogin = "13.1.0"
-    const val googlePlayServices = "20.1.0"
+    const val facebookLogin = "13.2.0"
+    const val googlePlayServices = "20.2.0"
     const val retrofit = "2.6.4"
     const val okhttp = "3.12.13"
     const val moshi = "1.13.0"
@@ -30,6 +30,8 @@ private object Versions {
     const val mockitoCore = "3.12.1"
     const val robolectric = "4.7.3"
 }
+
+const val ktLintVersion = "0.45.2"
 
 object Libs {
     const val annotations = "androidx.annotation:annotation:${Versions.annotationsLibrary}"