+++ plugins/src/main/kotlin/io/piano/android/ktlint/KtlintConfigPlugin.kt
@@ -1,5 +1,6 @@
 package io.piano.android.ktlint
 
+import io.piano.android.dependencies.ktLintVersion
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.gradle.kotlin.dsl.apply
@@ -11,6 +12,7 @@ class KtlintConfigPlugin : Plugin<Project> {
         apply<KtlintPlugin>()
         if (this != rootProject) {
             extensions.configure(KtlintExtension::class.java) {
+                version.set(ktLintVersion)
                 android.set(true)
             }
         }